Fix cmds() equivalent sending 2 instead of 4 bytes

+++ scripts_rend/socket-client.js
@@ -306,7 +306,7 @@ class SocketClient {
                 if (!tcoords) return;
 
                 buf.writeUInt16LE(tcoords.x, 1);
-                buf.writeUInt16LE(tcoords.y, 3);
+                buf.writeUInt32LE(tcoords.y, 3);
             break;
 
             case cl_cmds["CL_CMD_GIVE"]:
